+++ crates/vertigo-macro/src/lib.rs
@@ -20,7 +20,7 @@ use proc_macro::{Span, TokenStream};
 use quote::quote;
 
 use crate::{
-    api_access::api_access_inner,
+    api_access::api_access,
     bind::{bind_inner, bind_rc_inner, bind_spawn_inner},
     component::component_inner,
     css_parser::generate_css_string,
@@ -130,14 +130,8 @@ pub fn component(attrs: TokenStream, input: TokenStream) -> TokenStream {
 
 #[proc_macro]
 #[proc_macro_error]
-pub fn window(input: TokenStream) -> TokenStream {
-    api_access_inner("window", input)
-}
-
-#[proc_macro]
-#[proc_macro_error]
-pub fn document(input: TokenStream) -> TokenStream {
-    api_access_inner("document", input)
+pub fn js(input: TokenStream) -> TokenStream {
+    api_access(input)
 }
 
 fn convert_to_tokens(input: Result<TokenStream, String>) -> TokenStream {